Fixed shooter by seperating into two subsystems
